eval/requirements.txt
tiktoken>=0.7.0
scikit-learn>=1.3.0
networkx>=3.0
textstat>=0.7.3
pyyaml>=6.0
rich>=13.0
jsonschema>=4.0
pytest>=7.0
